Add a JSON loader in Angular
    In order to pass data between the server and the client without
breaking Django templating, a directive has been created which loads data from
a script tag with the type application/json. For security reason you must use
the JSONSafeEncoder for dumping the datas into the HTML in order to avoid XSS.

    It provides a json_angular template tag which inject JSON data in a safe
way and use an angular directive to retrieve it in the client.
    Usage is {% json_angular <object_to_dump> <id_of_the_object>%} where the
object will be dumped by horizon.utils.functions.JSONSafeEncoder and the id
could be a string or a value which will be interpolated by the django template
renderer.

Change-Id: Ica3dced5470b66934eaffd68dc85fcac3612c1c9
Implements: blueprint horizon-angular